+++ test/test_ads_fp_gen.py
@@ -9,10 +9,11 @@ import unittest
 from ase.build import fcc111, add_adsorbate
 from ase.data import atomic_numbers
 from ase.constraints import FixAtoms
-from catlearn.api.ase_atoms_api import database_to_list
+from catlearn.api.ase_atoms_api import database_to_list, images_connectivity
 from catlearn.fingerprint.adsorbate_prep import (autogen_info,
                                                  check_reconstructions,
-                                                 images_connectivity)
+                                                 connectivity2ads_index,
+                                                 slab_positions2ads_index)
 from catlearn.fingerprint.periodic_table_data import (get_radius,
                                                       default_catlearn_radius)
 from catlearn.fingerprint.setup import FeatureGenerator, default_fingerprinters
@@ -47,6 +48,9 @@ class TestAdsorbateFeatures(unittest.TestCase):
         print(str(len(images)) + ' training examples.')
         gen = FeatureGenerator(nprocs=1)
         train_fpv = default_fingerprinters(gen, 'adsorbates')
+        train_fpv += [gen.formal_charges,
+                      gen.ads_av,
+                      gen.ads_sum]
         matrix = gen.return_vec(images, train_fpv)
         labels = gen.return_names(train_fpv)
         print(np.shape(matrix), type(matrix))
@@ -96,7 +100,7 @@ class TestAdsorbateFeatures(unittest.TestCase):
         self.assertTrue(len(labels) == np.shape(matrix)[1])
 
     def test_recontruction(self):
-        images = self.setup_atoms()
+        images = database_to_list('data/ads_example.db')
         slabs = []
         for atoms in images:
             slab = atoms.copy()
@@ -106,7 +110,10 @@ class TestAdsorbateFeatures(unittest.TestCase):
         slabs = images_connectivity(slabs)
         image_pairs = zip(images, slabs)
         reconstructed = check_reconstructions(image_pairs)
-
+        for i in range(len(images)):
+            species = images[i].info['key_value_pairs']['species']
+            slab_positions2ads_index(images[i], slabs[i], species)
+            connectivity2ads_index(images[i], species)
         self.assertTrue(len(reconstructed) == 0)